Add List of known Pointers
With arrays there are going to be pointers on the stack. For garbage
collecting purposes and general execution, the evaluator needs to keep
track of those. (For example pop must remove more than a single element
from the stack, if there is a array on the stack.) Pointers are
differentiated between heap an stack by their most significant bit (not
implemented yet).