Evaluate Arrays on the stack
Arrays on the stack begin by their length in bytes followed by their
elements. So the array length instruction pushes the stack address of
the start of the array on the stack, since then everything about the
array is known by a single number. The fact, that the number on the
stack is a address and not just a number is encoded by the list of
pointers, the evaluator keeps track of.