Print Arrays as Arrays and not just as Pointers
This makes it so, that print will print pointers as arrays, with each
element listed separately. Also fixes that arrays pointed before to the
first element and not to its length. The print mechanism has to be
changed, if there are going to be actually type infos during runtime for
print. (So that bools can be printed as 'true' or 'false' and not as 0
or 1 e.g.)