Fix Bug where Strings may have extra 0 at the End
If the length of a string is the multiple of 4 the instruction_converter
pushed an extra 0 onto the stack, since it tried to handle missing
characters at the end, which don't completely fit a word. This is fixed
by checking, if there are any remaining characters, after walking
through all chunks of size 4.