Fix Bug where Arrays thought they were shorter
Arrays calculated there byte width, by adding the byte width of all
their children, but this is only true, if each child fits perfectly into
words. This is not always true for strings. Since multiple strings in
arrays don't share available space in words. So the actual byte width is
the size of each element rounded up to the next full word.