Rename OpCode ArrayLength to CreateStackPointer
This better reflects the action, this op code/instruction fulfills. Also
now the argument is used sensible, since before this was just the array
length a second time, which nobody used and might actually not be known
at compile time. CreateStackPointer now takes an argument for offset to
which the pointer should point begining by itself. So an offset of 0
would point at itself and an offset of 1 at the element before the
pointer on the stack.