Use Register Values in n-dimensional Array
If a array contains both literal values and variable references or array
accesses, this would print out wrong values before. Now the pointers stored in
the registers are pushed directly onto the stack instead of pushing them a
second time on the stack and accessing them through a stack pointer.