Fix Binary Expressions in Array Literals
Binary expressions like string concat emit a pointer on to the stack,
but this pointer was removed from the instruction list before, since the
array literal tried to provide its own instruction for that. This is now
fixed, so that only other literals get their last instruction replaced
by the array.