Add Program as Result from InstructionConverter
Before instruction converter returned just a list of instructions, but
the program also bundles the stack as static memory into the result of
the instruction converter. Before literals like string and array
literals where translated into instructions and load immediates. Now
they are directly written into the stack before hand. This also
concludes the storing of array literals only at the start of the stack
and not anywhere in the stack. This ensures, that pointer to arrays are
always correct. Because function calls can move elements on the stack,
to insert their own return pointer, arrays could be corrupted by this.
But when arrays are stored at the beginning, they will not be moved
around on the stack.