Add BoundNode::for_statement()
This moves the artificially creation of for statements into the bound
node and away from the binder. This might make sense if all lot more
bound nodes are replaced with simpler bound nodes. But maybe that will
not happen. Because currently only the instruction converter has jump
instructions, which need relative or absolute targets. Both are hard to
get by in a tree structure like the bound nodes are.