Add Control Flow Analyzer
This allows detecting, if a function always returns with a value. This also
means, that there is no need, to add an empty return statement at the end of non
void functions. This currently only works with simple functions, that have
neither if statements nor while loops.

Also the evaluator currently doesn't work with functions, that return a value.