Make Control Flow Analyzer public available
Bound Jump Nodes are now correctly transformed into basic blocks, for
that every basic block must be accessible by a label index, since
multiple labels can point to the same block, those label indices are
stored in a vec in the basic block. Also made all members of a
BasicBlock public so that they can be later used in a debug output.