Bind Constructor Calls
Constructor calls are very similiar to function calls, but currently
these are two different processes in the binder. This may change in the
future. The difference is, that the constructor takes its arguments from
the fields in the struct, while a function from its declaration. Also a
constructor does not represent actual code, so that it will be converted
differently to instructions then function calls.