Add Type::StructReference
This allows for recursive structs, which will always have a reference to
themselves, since the type with fields is only needed for constructors
and field accesses. Such a recursive struct could no be constructed
right now, since you would need infinite constructor calls for it. This
gets only interesting if noneable types are added. Or if arrays of
structs, since while you could supply empty arrays to the constructor,
the type system would always need to hold a reference to the type and
its fields.