Fix directly calling Closures
This is a feature where a closure will not become an actual closure, and
just called directly. This did not work correctly before, since the type
of the closure was just kept. That type was later used to determine,
that there must be a closure, which needs to be decoded. This is mostly
fixed now. Only the this_type gets lost on the function, but this does
not seem to make any trouble.