Add Instruction::WriteToHeap
WriteToHeap allocates the amount of words given in the argument on the
heap and writes the values from the stack onto the heap. This allows for
an easy conversion of a stack value to a heap value. This will be used
e.g. to box integers when they are wrapped in a noneable type.