Convert any type via bind_conversion
This handles all boxing in one place now. In the future type any may
also be handled by this, so that the system calls would not need a
special handling. But this would need some changing, since then type any
would have to be a heap allocated object, since you could otherwise not
easily detect if you have a value or a type identifier on the stack.
Since this are quite some changes they are not done yet.