Allocate Structs, Arrays and Closures on Heap
Before those were allocated on the stack, but if there was a loop or as
function, that was called multiple times, this meant that they were
allocated again and again in the same space, which of course is quite
wrong. So now everything, that can be modified during runtime is
allocated on the heap and only string literals are still on the stack.