Start implementing type boxing
Type boxing describes turning the type information for a value into a
runtime readable format, so that the program can do different things
depending on the type. This already exists in system calls like print or
to$string, which turn any type into a human readable format. This will
unify the place where type boxing happens and will later also allow the
user to write code that interacts with values of any type.