Add Instruction::TypeIdentifierEquals
This instruction is used in casts, to check if the types of two
expressions are the same. It also leaves the value that results after
the cast on the stack. So if the comparison fails it would push a none
pointer onto the stack. And if it succeeds it pushes the value inside
the any value on to the stack. The result is not necessary equal to a
Noneable value, since the value inside any does not have to be a
pointer.