Restructure dependency_resolver.rs
It makes more sense, that the import nodes are not replaced by a
dependency resolver, since there is a lot of shared functionality with a
binder. In the future the binder will bind the import functions, the
arguments and their values. Dependency resolver will hold all possible
import functions and other needed data types.