Add versions of all function to compile libraries
The binder has only a flag because currently the only difference is
inserting a call to main if it is not a library. In the future this will
probably change, since libraries and presentations have completely
different top level statements. In the instruction_converter this is
already done, since in the near future we also want to export the
symbols inside a library and not only the instructions to execute the
symbols.