Add variable_id to FunctionDeclarationBody
Variable id is the actual variable id that a function has, since
currently there are no anonymous functions, all functions have one. The
difference to the function id is, that a function id is more like the
label of the function, which will later be resolved to a address in the
code, while the variable id will be resolved to a register in which that
function is referenced.