Fix bug where compiler memory overflowed sometimes
If the parser was in a while loop, which would only end if a certain
token would appear, this could mean, that it never stops, since until
the eoi token, there might be no token needed. Than it would loop
endlessly on the eoi token. So now the eoi token is checked against.