Fix bug where member functions had the wrong name
Inside the compiler there are only normal functions and no member
functions. Which means any member function will be turned into a normal
function with `this` as first argument. To ensure, that there is no name
clashing they will be prefixed with the struct name and a colon colon,
to ensure, that the user cannot accidentally call this function. This
has now been correctly implemented.