Split program into functions and startup
Since the necessary setup for functions was lost when are library was
imported, since the startup code of the library was lost, the startup
code is now its own entity. Which means that all libraries just collect
startup code from other libraries and only the program in the end will
insert all collected startup code in its program code. This also mean
that the setup code now cannot be checked in the -dbp flag, since this
only shows bound nodes and not instructions. The startup code needs to
be instructions, since a bound node has string literals, with lifetimes,
which a library cannot have. In the future a dissassembler for the
instruction code might make sense, but is probably also a bigger
project. (Just like imports..)