Prepare importing the same library multiple times
Since there is a valid reason to import the same library in multiple
other libraries, the same libraries can be imported multiple times. Of
course it is expected, that types or functions from that library are the
same across all libraries. This commit prepares, that all imported
libraries can be searched, if they contain a certain library. If they
do, they do not be loaded again.