Add a new kind of Closure
This closure holds a pointer to the actual code and not just a variable
(like FunctionId does). This allows to use closures with constants,
since functions from extern libraries are sometimes converted into
constants instead of variables. This will also happen more often in the
future for a faster runtime.