Allow function overshadowing
Also edit error message a bit to be clearer if a field on a struct is
declared multiple times. The user can now declare functions with the
same name as normal fields. This will overshadow the field from the
outside and overshadow the function from inside the struct. This will
also allow overloading in the future (not yet implemented).