Move toString system call into binder
Before the instruction converter inserted the necessary instructions to
convert a value to a string, but now the binder does this automatically
and will also call the appropiate functions if available. This also
means, that these functions will not be called for types like any or
noneable. For these types there need to be other solutions.