Rework diagnostics messages
Before diagnostics were just strings, but this does expect, that all
type names are known at the time an error is reported. But with generic
types this will change, since every constructor can generate a new type.
So this fixes this, by only supplying the type names, after all types
are known and diagnostics only getting converted to string, when they
are actually displayed.