Fix loss of information
The compiler lost information of compiler functions like $toString in
structs, if the struct was applied to a generic struct. This meant, that
arrays of structs were printed very ugly. This is fixed now by making
the $toString (and all compiler functions in general) be known even to
just the references of a struct.