Split Type::Integer and Type::UnsignedInteger
Since there are multiple types of integer, they need to be seperate
types. But integers are also very similiar in a lot of ways and also
want to be converted into each other. So Type::Integer(integer_type) now
has the integer_type which holds information of the sign and byte size
of the integer.

The other type that was introduced is the Type::IntegerLiteral, which
are integer literals, since they have no explicit type information, and
need to be easily converted into differen integer types.