Fix label overlap
Before it happened, that the main function had a label which was already
used in a library function, this lead to wrong jumps in the program and
stack overflows, because the main function kept being executed forever.
This was fixed by this, but it might be, that there is an unused label
between libraries and program. But that shouldn't be too bad.