Register libraries as constants instead of variables
Before variables where declared as variables, which led to wrong
behaivour, since in many places the program did not expect any variables
to be declared at certain points. This is now fixed by actually creating
values for libraries which allows setting them as constants. This also
means that `protected_registers` in evaluator should always be zero.
This will probably removed in a future commit.